feat(plugins): change file names
