+++ packages/plugin-session-replay-browser/jest.config.js
