+++ packages/plugin-session-replay-browser/package.json
