+++ packages/plugin-session-replay-browser/rollup.config.js
