+++ packages/plugin-session-replay-browser/src/constants.ts
