+++ packages/plugin-session-replay-browser/src/helpers.ts
