+++ packages/plugin-session-replay-browser/src/index.ts
