+++ packages/plugin-session-replay-browser/test/session-replay.test.ts
