+++ packages/plugin-session-replay-browser/tsconfig.es5.json
