+++ packages/plugin-session-replay-browser/tsconfig.esm.json
