+++ packages/plugin-session-replay-browser/tsconfig.json
