--- packages/plugin-session-replay/src/typings/session-replay.ts
@@ -1,16 +1,0 @@
-export type Events = string[];
-
-export interface SessionReplayContext {
-  events: Events;
-  sequenceId: number;
-  attempts: number;
-  timeout: number;
-  sessionId: number;
-}
-
-export interface IDBStore {
-  [sessionId: number]: {
-    events: Events;
-    sequenceId: number;
-  };
-}
